Self delete

+++ crates/uv/src/commands/tool/uninstall.rs
@@ -180,6 +180,9 @@ async fn uninstall_tool(
     // Remove the tool itself.
     tools.remove_environment(name)?;
 
+    #[cfg(windows)]
+    let itself = std::env::current_exe().ok();
+
     // Remove the tool's entrypoints.
     let entrypoints = receipt.entrypoints();
     for entrypoint in entrypoints {
@@ -187,6 +190,15 @@ async fn uninstall_tool(
             "Removing executable: {}",
             entrypoint.install_path.user_display()
         );
+
+        #[cfg(windows)]
+        if itself.as_ref().is_some_and(|itself| {
+            std::path::absolute(&entrypoint.install_path).is_ok_and(|target| *itself == target)
+        }) {
+            self_replace::self_delete()?;
+            continue;
+        }
+
         match fs_err::tokio::remove_file(&entrypoint.install_path).await {
             Ok(()) => {}
             Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
